+++ cmd/registry/main.go
@@ -12,10 +12,10 @@ import (
 	"github.com/gorilla/handlers"
 	"github.com/yvasiyarov/gorelic"
 
-	"github.com/docker/distribution"
 	_ "github.com/docker/distribution/auth/silly"
 	_ "github.com/docker/distribution/auth/token"
 	"github.com/docker/distribution/configuration"
+	"github.com/docker/distribution/registry"
 	_ "github.com/docker/distribution/storagedriver/filesystem"
 	_ "github.com/docker/distribution/storagedriver/inmemory"
 )
