+++ registry/api_test.go
