+++ registry/app.go
