+++ registry/app_test.go
