+++ registry/context.go
