+++ registry/images.go
