+++ registry/tokens_test.go
